requirements.txt
alembic==1.10.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
flask-change-password==0.0.7
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Share==0.1.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
Pillow==9.5.0
python-dotenv==0.21.1
requests==2.28.2
SQLAlchemy==2.0.3
typing_extensions==4.4.0
urllib3==1.26.15
Werkzeug==2.2.2
WTForms==3.0.1
